1,74.0

# -----BEGIN PGP SIGNATURE-----
#
# iQIzBAABCAAdFiEEGu1npAng4grd+K9FGB/JDRU5PCAFAmOQ83MACgkQGB/JDRU5
# PCAqZA/+J5Ftn0oFpPrSBAGuzk00xoBeK2be0AKNbU3PGwY039sNXxQaeZ7xWLuD
# oCIQdYVn7ofEDZUKaORntfrxlxWJGAzKQt3TS5T+bBSI0eoGiAu+YOrEFdbh5XE6
# TtN32yneK9KyNL16oRHvH9jm5xJTFahayn8DzYnTJA3bWBu3/Pl02lvo6zMv1Gcj
# x/xadYFIpybEUxXduHp3ZwAboqV9ntteS0uObrG9B+VFKlmcX5bTAf6ZJMUzPTOi
# OU69Szz4G6nU4QBflN1JU6f505UMEQNvqfnYSqsAgXV4CIQxm6EopEhYcQn5tQCu
# pRvjDF0Q3UB8JswepFIBJWXbxWO7nEOBeDX6z6BTO37W3+hs7Kk8S8/JqPaJG9Y0
# xXXrFTIspD1FdDoLQQ47nlIJ4JbpwOopC6JHk4La6EnA/NYRun6R/A9sRFEEJ6ns
# gp4pZPweD5QHC38hVhPOFmrvNK91FNxIpHUCRL5LEzmDNy6qV7PTlrOcOsFWjo8T
# dD3H4jUQdKHctw30yoj2QR0hPCEkhD3SyNPhnX5HZDbwTb7EUtBrXIv1Nbx4u0Zk
# gtxp3LbjG7/S6gNeIh3NTTz/cNpUHvnPQQR3zbI59HK54Nrw6k54cYMdsDo+chyF
# XCHzgWM6zQlIftG6ToVf8JLM/9gHAlOPvwbxy8ShC0RF7w/vEdA=
# =4te4
# -----END PGP SIGNATURE-----
# gpg: Signature made Wed Dec  7 21:11:31 2022
# gpg:                using RSA key 1AED67A409E0E20ADDF8AF45181FC90D15393C20
# gpg: Can't check signature: No public key

# Conflicts:
#	.github/classifier.json
#	.vscode/notebooks/endgame.github-issues
#	.vscode/notebooks/my-endgame.github-issues
#	.vscode/notebooks/verification.github-issues
#	.vscode/settings.json
#	extensions/git/package.json
#	extensions/git/src/commands.ts
#	extensions/git/src/main.ts
#	extensions/git/src/model.ts
#	extensions/git/yarn.lock
#	extensions/github-authentication/package.json
#	extensions/github-authentication/yarn.lock
#	extensions/html-language-features/package.json
#	extensions/json-language-features/package.json
#	extensions/markdown-language-features/package.json
#	extensions/markdown-language-features/yarn.lock
#	extensions/media-preview/package.json
#	extensions/microsoft-authentication/extension-browser.webpack.config.js
#	extensions/microsoft-authentication/package.json
#	extensions/microsoft-authentication/src/AADHelper.ts
#	extensions/microsoft-authentication/src/browser/authServer.ts
#	extensions/microsoft-authentication/src/browser/buffer.ts
#	extensions/microsoft-authentication/src/node/authServer.ts
#	extensions/microsoft-authentication/src/utils.ts
#	extensions/microsoft-authentication/yarn.lock
#	extensions/simple-browser/package.json
#	extensions/simple-browser/yarn.lock
#	extensions/typescript-language-features/package.json
#	extensions/typescript-language-features/src/experimentTelemetryReporter.ts
#	extensions/typescript-language-features/src/experimentationService.ts
#	extensions/typescript-language-features/src/languageFeatures/quickFix.ts
#	package.json
#	product.json
#	remote/package.json
#	remote/web/package.json
#	remote/web/yarn.lock
#	src/vs/code/electron-main/app.ts
#	src/vs/code/node/cliProcessMain.ts
#	src/vs/editor/contrib/codeAction/browser/codeActionWidget.ts
#	src/vs/editor/contrib/inPlaceReplace/browser/inPlaceReplace.css
#	src/vs/editor/standalone/browser/standaloneServices.ts
#	src/vs/platform/externalServices/common/marketplace.ts
#	src/vs/platform/keybinding/common/abstractKeybindingService.ts
#	src/vs/platform/telemetry/common/serverTelemetryService.ts
#	src/vs/platform/telemetry/common/telemetry.ts
#	src/vs/platform/telemetry/common/telemetryService.ts
#	src/vs/platform/telemetry/common/telemetryUtils.ts
#	src/vs/platform/telemetry/node/customEndpointTelemetryService.ts
#	src/vs/platform/telemetry/test/browser/telemetryService.test.ts
#	src/vs/platform/telemetry/test/common/telemetryLogAppender.test.ts
#	src/vs/platform/userDataSync/common/abstractSynchronizer.ts
#	src/vs/platform/userDataSync/common/userDataAutoSyncService.ts
#	src/vs/workbench/api/browser/mainThreadTelemetry.ts
#	src/vs/workbench/api/common/extHost.protocol.ts
#	src/vs/workbench/api/common/extHostTelemetry.ts
#	src/vs/workbench/api/common/extensionHostMain.ts
#	src/vs/workbench/api/test/browser/extHostApiCommands.test.ts
#	src/vs/workbench/api/test/browser/extHostLanguageFeatures.test.ts
#	src/vs/workbench/api/test/browser/extHostTelemetry.test.ts
#	src/vs/workbench/browser/parts/editor/editorGroupView.ts
#	src/vs/workbench/browser/parts/editor/titleControl.ts
#	src/vs/workbench/contrib/debug/browser/breakpointsView.ts
#	src/vs/workbench/contrib/debug/browser/debugSession.ts
#	src/vs/workbench/contrib/editSessions/browser/editSessions.contribution.ts
#	src/vs/workbench/contrib/extensions/browser/dynamicWorkspaceRecommendations.ts
#	src/vs/workbench/contrib/interactive/browser/interactiveEditor.ts
#	src/vs/workbench/contrib/localization/electron-sandbox/localization.contribution.ts
#	src/vs/workbench/contrib/logs/common/logs.contribution.ts
#	src/vs/workbench/contrib/mergeEditor/browser/mergeEditorInputModel.ts
#	src/vs/workbench/contrib/mergeEditor/test/browser/model.test.ts
#	src/vs/workbench/contrib/notebook/browser/notebookEditor.ts
#	src/vs/workbench/contrib/search/browser/searchView.ts
#	src/vs/workbench/contrib/search/test/electron-browser/textsearch.perf.integrationTest.ts
#	src/vs/workbench/contrib/tags/electron-sandbox/workspaceTags.ts
#	src/vs/workbench/contrib/telemetry/browser/telemetry.contribution.ts
#	src/vs/workbench/contrib/terminal/browser/xterm/quickFixAddon.ts
#	src/vs/workbench/contrib/testing/browser/testingExplorerView.ts
#	src/vs/workbench/services/extensions/common/extensionHostManager.ts
#	src/vs/workbench/services/languageDetection/browser/languageDetectionWorkerServiceImpl.ts
#	src/vs/workbench/services/telemetry/browser/telemetryService.ts
#	src/vs/workbench/services/telemetry/electron-sandbox/telemetryService.ts
#	src/vs/workbench/services/timer/browser/timerService.ts
#	src/vscode-dts/vscode.proposed.telemetryLogger.d.ts
#	yarn.lock

+++ src/vs/platform/windows/electron-main/windowImpl.ts
@@ -41,8 +41,8 @@ import { Color } from 'vs/base/common/color';
 import { IPolicyService } from 'vs/platform/policy/common/policy';
 import { IUserDataProfile } from 'vs/platform/userDataProfile/common/userDataProfile';
 import { IStateMainService } from 'vs/platform/state/electron-main/state';
-import product from 'vs/platform/product/common/product';
 import { IUserDataProfilesMainService } from 'vs/platform/userDataProfile/electron-main/userDataProfile';
+import { INativeHostMainService } from 'vs/platform/native/electron-main/nativeHostMainService';
 
 export interface IWindowCreationOptions {
 	state: IWindowState;
@@ -85,6 +85,8 @@ export class CodeWindow extends Disposable implements ICodeWindow {
 
 	private static readonly windowControlHeightStateStorageKey = 'windowControlHeight';
 
+	private static sandboxState: boolean | undefined = undefined;
+
 	//#region Events
 
 	private readonly _onWillLoad = this._register(new Emitter<ILoadEvent>());
@@ -120,7 +122,16 @@ export class CodeWindow extends Disposable implements ICodeWindow {
 
 	get openedWorkspace(): IWorkspaceIdentifier | ISingleFolderWorkspaceIdentifier | undefined { return this._config?.workspace; }
 
-	get profile(): IUserDataProfile | undefined { return this.config ? this.userDataProfilesService.getOrSetProfileForWorkspace(this.config.workspace ?? 'empty-window', this.userDataProfilesService.profiles.find(profile => profile.id === this.config?.profiles.profile.id) ?? this.userDataProfilesService.defaultProfile) : undefined; }
+	get profile(): IUserDataProfile | undefined {
+		if (!this.config) {
+			return undefined;
+		}
+		const profile = this.userDataProfilesService.profiles.find(profile => profile.id === this.config?.profiles.profile.id);
+		if (this.isExtensionDevelopmentHost && profile) {
+			return profile;
+		}
+		return this.userDataProfilesService.getOrSetProfileForWorkspace(this.config.workspace ?? 'empty-window', profile ?? this.userDataProfilesService.defaultProfile);
+	}
 
 	get remoteAuthority(): string | undefined { return this._config?.remoteAuthority; }
 
@@ -178,7 +189,8 @@ export class CodeWindow extends Disposable implements ICodeWindow {
 		@IProductService private readonly productService: IProductService,
 		@IProtocolMainService private readonly protocolMainService: IProtocolMainService,
 		@IWindowsMainService private readonly windowsMainService: IWindowsMainService,
-		@IStateMainService private readonly stateMainService: IStateMainService
+		@IStateMainService private readonly stateMainService: IStateMainService,
+		@INativeHostMainService private readonly nativeHostMainService: INativeHostMainService
 	) {
 		super();
 
@@ -193,13 +205,21 @@ export class CodeWindow extends Disposable implements ICodeWindow {
 			// after the call to maximize/fullscreen (see below)
 			const isFullscreenOrMaximized = (this.windowState.mode === WindowMode.Maximized || this.windowState.mode === WindowMode.Fullscreen);
 
+			if (typeof CodeWindow.sandboxState === 'undefined') {
+				// we should only check this once so that we do not end up
+				// with some windows in sandbox mode and some not!
+				CodeWindow.sandboxState = this.stateMainService.getItem<boolean>('window.experimental.useSandbox', false);
+			}
+
 			const windowSettings = this.configurationService.getValue<IWindowSettings | undefined>('window');
 
 			let useSandbox = false;
 			if (typeof windowSettings?.experimental?.useSandbox === 'boolean') {
 				useSandbox = windowSettings.experimental.useSandbox;
+			} else if (this.productService.quality === 'stable' && CodeWindow.sandboxState) {
+				useSandbox = true;
 			} else {
-				useSandbox = typeof product.quality === 'string' && product.quality !== 'stable';
+				useSandbox = typeof this.productService.quality === 'string' && this.productService.quality !== 'stable';
 			}
 
 			const options: BrowserWindowConstructorOptions & { experimentalDarkMode: boolean } = {
@@ -213,7 +233,7 @@ export class CodeWindow extends Disposable implements ICodeWindow {
 				show: !isFullscreenOrMaximized, // reduce flicker by showing later
 				title: this.productService.nameLong,
 				webPreferences: {
-					preload: FileAccess.asFileUri('vs/base/parts/sandbox/electron-browser/preload.js', require).fsPath,
+					preload: FileAccess.asFileUri('vs/base/parts/sandbox/electron-browser/preload.js').fsPath,
 					additionalArguments: [`--vscode-window-config=${this.configObjectUrl.resource.toString()}`],
 					v8CacheOptions: this.environmentMainService.useCodeCache ? 'bypassHeatCheck' : 'none',
 					enableWebSQL: false,
@@ -692,34 +712,65 @@ export class CodeWindow extends Disposable implements ICodeWindow {
 
 				// Process gone
 				else if (type === WindowError.PROCESS_GONE) {
-					let message: string;
-					if (!details) {
-						message = localize('appGone', "The window terminated unexpectedly");
-					} else {
-						message = localize('appGoneDetails', "The window terminated unexpectedly (reason: '{0}', code: '{1}')", details.reason, details.exitCode ?? '<unknown>');
+
+					// Windows: running as admin with AppLocker enabled is unsupported
+					//          when sandbox: true.
+					//          we cannot detect AppLocker use currently, but make a
+					//          guess based on the reason and exit code.
+					if (isWindows && details?.reason === 'launch-failed' && details.exitCode === 18 && await this.nativeHostMainService.isAdmin(undefined)) {
+						await this.handleWindowsAdminCrash(details);
 					}
 
-					// Show Dialog
-					const result = await this.dialogMainService.showMessageBox({
-						title: this.productService.nameLong,
-						type: 'warning',
-						buttons: [
-							mnemonicButtonLabel(localize({ key: 'reopen', comment: ['&& denotes a mnemonic'] }, "&&Reopen")),
-							mnemonicButtonLabel(localize({ key: 'close', comment: ['&& denotes a mnemonic'] }, "&&Close"))
-						],
-						message,
-						detail: localize('appGoneDetail', "We are sorry for the inconvenience. You can reopen the window to continue where you left off."),
-						noLink: true,
-						defaultId: 0,
-						checkboxLabel: this._config?.workspace ? localize('doNotRestoreEditors', "Don't restore editors") : undefined
-					}, this._win);
+					// Any other crash: offer to restart
+					else {
+						let message: string;
+						if (!details) {
+							message = localize('appGone', "The window terminated unexpectedly");
+						} else {
+							message = localize('appGoneDetails', "The window terminated unexpectedly (reason: '{0}', code: '{1}')", details.reason, details.exitCode ?? '<unknown>');
+						}
 
-					// Handle choice
-					const reopen = result.response === 0;
-					await this.destroyWindow(reopen, result.checkboxChecked);
+						// Show Dialog
+						const result = await this.dialogMainService.showMessageBox({
+							title: this.productService.nameLong,
+							type: 'warning',
+							buttons: [
+								mnemonicButtonLabel(localize({ key: 'reopen', comment: ['&& denotes a mnemonic'] }, "&&Reopen")),
+								mnemonicButtonLabel(localize({ key: 'close', comment: ['&& denotes a mnemonic'] }, "&&Close"))
+							],
+							message,
+							detail: localize('appGoneDetail', "We are sorry for the inconvenience. You can reopen the window to continue where you left off."),
+							noLink: true,
+							defaultId: 0,
+							checkboxLabel: this._config?.workspace ? localize('doNotRestoreEditors', "Don't restore editors") : undefined
+						}, this._win);
+
+						// Handle choice
+						const reopen = result.response === 0;
+						await this.destroyWindow(reopen, result.checkboxChecked);
+					}
 				}
 				break;
 		}
+	}
+
+	private async handleWindowsAdminCrash(details: { reason: string; exitCode: number }) {
+
+		// Inform user
+		await this.dialogMainService.showMessageBox({
+			title: this.productService.nameLong,
+			type: 'error',
+			buttons: [
+				mnemonicButtonLabel(localize({ key: 'close', comment: ['&& denotes a mnemonic'] }, "&&Close"))
+			],
+			message: localize('appGoneAdminMessage', "Running as administrator is not supported"),
+			detail: localize('appGoneAdminDetail', "Please try again without administrator privileges.", this.productService.nameLong),
+			noLink: true,
+			defaultId: 0
+		}, this._win);
+
+		// Exit
+		await this.destroyWindow(false, false);
 	}
 
 	private async destroyWindow(reopen: boolean, skipRestoreEditors: boolean): Promise<void> {
@@ -860,7 +911,7 @@ export class CodeWindow extends Disposable implements ICodeWindow {
 		this.readyState = ReadyState.NAVIGATING;
 
 		// Load URL
-		this._win.loadURL(FileAccess.asBrowserUri(`vs/code/electron-sandbox/workbench/workbench${this.environmentMainService.isBuilt ? '' : '-dev'}.html`, require).toString(true));
+		this._win.loadURL(FileAccess.asBrowserUri(`vs/code/electron-sandbox/workbench/workbench${this.environmentMainService.isBuilt ? '' : '-dev'}.html`).toString(true));
 
 		// Remember that we did load
 		const wasLoaded = this.wasLoaded;
@@ -961,6 +1012,7 @@ export class CodeWindow extends Disposable implements ICodeWindow {
 			all: this.userDataProfilesService.profiles,
 			profile: this.profile || this.userDataProfilesService.defaultProfile
 		};
+		configuration.logLevel = this.logService.getLevel();
 
 		// Load config
 		this.load(configuration, { isReload: true, disableExtensions: cli?.['disable-extensions'] });
